Updates to search for block by id
