+++ lib/blocks/widgets/block_slider/block_view.dart
@@ -1,12 +1,9 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_annulus/blocks/sections/block_details_drawer.dart';
-import 'package:flutter_annulus/blocks/sections/block_mobile_details.dart';
 import 'package:flutter_annulus/blocks/utils/utils.dart';
 import 'package:flutter_annulus/shared/theme.dart';
+import 'package:flutter_annulus/shared/utils/nav_utils.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
-import 'package:modal_side_sheet/modal_side_sheet.dart';
 import 'package:responsive_framework/responsive_breakpoints.dart';
-import 'package:vrouter/vrouter.dart';
 
 import '../../../shared/providers/app_theme_provider.dart';
 import '../../../shared/utils/theme_color.dart';
@@ -25,7 +22,6 @@ class BlockView extends ConsumerWidget {
   @override
   Widget build(BuildContext context, WidgetRef ref) {
     final isMobile = ResponsiveBreakpoints.of(context).equals(MOBILE);
-    final isDesktop = ResponsiveBreakpoints.of(context).equals(DESKTOP);
 
     final colorTheme = ref.watch(appThemeColorProvider);
     return Container(
@@ -51,15 +47,10 @@ class BlockView extends ConsumerWidget {
         data: (Block block) {
           return TextButton(
               onPressed: () {
-                isDesktop
-                    ? showModalSideSheet(
-                        context: context,
-                        ignoreAppBar: false,
-                        width: 640,
-                        barrierColor: getSelectedColor(colorTheme, 0xFFFEFEFE, 0xFF353739).withOpacity(0.64),
-                        barrierDismissible: true,
-                        body: BlockDetailsDrawer(block: block))
-                    : context.vRouter.to(BlockTabBarMobileView.blockDetailsPath(block.header));
+                goToBlockDetails(
+                  context: context,
+                  block: block,
+                );
               },
               style: ButtonStyle(
                 foregroundColor: MaterialStateProperty.all<Color>(Colors.blue),
